Removes local zustand business logic for storing data via localStorage
